requirements.txt
django==1.10.1
ujson==1.35
pymongo==3.3.0
Wand==0.4.3
python-magic==0.4.12
pycrypto==2.6.1
redis==2.10.5